去除编译告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@30363 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/ixm/ixmIndexKey.cpp
@@ -415,7 +415,7 @@ namespace engine
             string keyName ;
             if ( isKeepKeyName )
             {
-               if ( i == arrElePos )
+               if ( (INT32)i == arrElePos )
                {
                   SDB_ASSERT( arrNameLen >= 0, "invalid arrNameLen" ) ;
                   keyName = string( fieldNames[i], arrNameLen ) ;
